[JS Project] JS Function #9
I'd like to make functions for login page.

Firstly, the users input their own name or user id.
If there no any informations about the user in a localStorage, then after clicking the "LOGIN" button,  the information about the user is saved in the localStorage.
However, if the informations exist in the localStorage, then just fetch the data from there and the login page will be disappeared and the user can use their dailiy planner.

I am making this functions with JS grammar studying... (22. 03. 11~ )